refactor: Move and refactor InMemoryKeyBindingRepository
